refactor beta rowset implementation into rowset, remove the inheritant hierarchy.
